Write unit tests for Blockchain sync, fix some minor frontend bugs
I wrote a few high-level (behaviour) tests for the Blockchain sync function,
using mocking that check that all the correct logic is impemented and correct
functions are called.
Fixed a bug in js with setInterval that prevented latest blocks from showing up
on index page load.
Added a link for the wallet bootstrap.zip file.