Bugfix for stopped syncing in between 99% and 100%
Bug occures when syncing stops in between 99% and 100%, because in that window regular sync stops and rpc sync kicks in,
and we start the sync again the next time. The next time sync starts, it will try to sync with regular sync method,
which needs dat field in mongo to know from which block to start when syncing via rpc, we don't insert dat field into mongo,
and it causes null error. Easiest solution is to start the resync directly with rpc method, it will be a little slower but
it wont cause null error.