Bugfix for rpc_sync, inserting raw transaction vout index into Mongo
VoutSerializer.to_database_rpc in serializers.py was causing AttributeError? because vout.n was undefined,
instead of forwarding True if rpc_sync is used for Mongo syncing in flush_cache, all raw transactions
are forwarded. Now in serializers, those raw transactions can be used to get the index of vout.