Hotfix for block and transaction confirmations added
Calculations for block and transaction confirmations has been changed to check if it is from main or side chain.
Reason: If it is from main chain we are calculating it in code: bestblock height - ourblock height, to optimise the server, if
it is from side chain we get it from directly from client.